Extend status with "private" and "future", for admin
