+++ layers/WPSchema/packages/customposts/src/Enums/CustomPostStatus.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPWPSchema\CustomPosts\Enums;
+
+class CustomPostStatus
+{
+    public const FUTURE = 'future';
+    public const PRIVATE = 'private';
+}
